gnulib: store result of bootstrap in git
There is no need to run bootstrap every time a clean git checkout is
done as the result will be always the same.  We only need to run it
when there is some change to gnulib modules or when we update to new
gnulib or there is a change to the build system.  All of the steps
above are done as by developers and therefore it can be part of git.

The benefits are that from this point nothing will modify source dir
when building libvirt, the configure time will be way quicker and it
helps with the planned adoption of Meson as our build system.

The drawbacks are that developers will have to regenerate these files
when doing changes mentioned above and there is a significant amount
of files and code lines part of our git checkout.  However, with the
ongoing effort to eliminate gnulib in favor of GLib all of the gnulib
code will be eventually dropped.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>